+++ src/main/java/com/mz/sistema/gestao/escolar/servico/ProfessorTurmaServico.java
@@ -16,9 +16,9 @@ public interface ProfessorTurmaServico {
 	public void salvar(ProfessorTurma professorTurma);
 	public void excluir(ProfessorTurma professorTurma);
 	public List<ProfessorTurma> obterTurmasDoProfessorPorEscolaAno(Long idEscolaProfessor, Integer ANO);
-	public List<ProfessorTurma> obterProfessorTurmaPorTurma(Integer idTurma);
-	public List<ProfessorTurma> obterProfessorTurmaPorTurmaOrdenarPorIdDisciplina(Integer idTurma);
-	public ProfessorTurma obterProfessorTurmaPorIdTurmarPorIdDisciplina(Integer idTurma, Integer idDisciplina);
-	public List<ProfessorTurma> obterProfessorTurmaPorTurmaOrdernarPorNome(Integer idTurma);
+	public List<ProfessorTurma> obterProfessorTurmaPorTurma(Long idTurma);
+	public List<ProfessorTurma> obterProfessorTurmaPorTurmaOrdenarPorIdDisciplina(Long idTurma);
+	public ProfessorTurma obterProfessorTurmaPorIdTurmarPorIdDisciplina(Long idTurma, Long idDisciplina);
+	public List<ProfessorTurma> obterProfessorTurmaPorTurmaOrdernarPorNome(Long idTurma);
 	
 }
